fix blocknumber

+++ worker.go
@@ -64,7 +64,7 @@ func (w *Worker) ClusterWithdrawal(id gtypes.WorkId, val int64, untilFinalized b
 func (w *Worker) ClusterUnmortgage(clusterID uint64, id uint64, untilFinalized bool) error {
 	runtimeCall := weteeworker.MakeClusterUnmortgageCall(
 		clusterID,
-		uint32(id),
+		id,
 	)
 
 	return w.Client.SignAndSubmit(w.Signer, runtimeCall, untilFinalized)
